% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-utils.R
\name{append_list_replace}
\alias{append_list_replace}
\title{append replace list}
\usage{
append_list_replace(list0, ...)
}
\arguments{
\item{list0}{the list to append to (w/ replacement)}

\item{\dots}{additional lists or .yaml filepaths to add to or replace elements in list0}
}
\value{
the merged list with later elements in \dots replacing earlier ones
}
\description{
merge multiple lists with replacement
}
\details{
order matters for ... arguments. The LAST argument will overwrite anything that proceeds it.
That means that \code{append_list_replace(list(dog='Larry'), list(dog='Cindy'))}
will use \code{dog='Cindy'}.
}
