% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-utils.R
\name{as.attr_file}
\alias{as.attr_file}
\title{create name for attributes file if one doesn't exist}
\usage{
as.attr_file(filename)
}
\arguments{
\item{filename}{the name of the files that this attribute file will be based on}
}
\value{
a filepath for an attributes file
}
\description{
create attribute file name
}
