% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-utils.R
\name{as.attr_list}
\alias{as.attr_list}
\title{read and format attributes from delimited file or data.frame}
\usage{
as.attr_list(x)
}
\arguments{
\item{x}{a file name for attributes file or a data.frame from \code{\link{read_attr_file}}}
}
\value{
a list of attributes and values for rendering
}
\description{
formats an attribute table into a list appropriate for metadata rendering
}
\keyword{internal}

