% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-utils.R
\name{write_attr_file}
\alias{read_attr_file}
\alias{write_attr_file}
\title{write a file formatted as attributes}
\usage{
write_attr_file(x, file)

read_attr_file(attr.file)
}
\arguments{
\item{x}{the data to be written to file}

\item{file}{a filepath to a file that is comma or tab delimited}

\item{attr.file}{a filepath to a file that is comma or tab delimited}
}
\value{
the data.frame of formatted attributes
}
\description{
write external file for attributes

read in external file for attributes
}

