% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-utils.R
\name{attr_names}
\alias{attr_names}
\title{read attributes from a recognized file type}
\usage{
attr_names(x)
}
\arguments{
\item{x}{the full path for the file(s) or an object}
}
\value{
the attributes as a character vector
}
\description{
return just the header names (attributes) for delimited files or shapefiles
}
\examples{
# attribute names from a shapefile:
attr_names(system.file(package='meddle','extdata','example_shapefile'))

# attribute names from a Spatial object:
sf <- read_data(system.file(package='meddle','extdata','example_shapefile'))
attr_names(sf)
}
