% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_skeleton.R
\name{attribute_skeleton.character}
\alias{attribute_skeleton.character}
\title{attribute skeleton for an assumed file or files}
\usage{
\method{attribute_skeleton}{character}(object, attr.file, ...)
}
\description{
will attempt to figure out the content, read in the file(s) and then
pass to appropriate method
}
\keyword{internal}

