% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_skeleton.R
\name{attribute_skeleton}
\alias{attribute_skeleton}
\title{create a skeleton file (or safely append to existing) for attributes}
\usage{
attribute_skeleton(object, attr.file, ...)
}
\arguments{
\item{object}{a regognized R object or vector of file names that can be read in to a recognized object class}

\item{attr.file}{a file to write the attribute skeleton to}

\item{\dots}{additional parameters passed to methods}
}
\description{
create skeleton from object
}

