% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-utils.R
\name{environment_metadata}
\alias{environment_metadata}
\title{get metadata for computing environment}
\usage{
environment_metadata()
}
\description{
environment metadata returned as a string
}
\details{
create a string from sessionInfo() that can be used 
for `Native Data Set Environment`
}
