% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-utils.R
\name{eval_content}
\alias{eval_content}
\title{evaluate content blocked within r code}
\usage{
eval_content(x)
}
\arguments{
\item{x}{any R object}
}
\value{
the return value from evaluating the function, or the raw input (unaltered)
if there is no function to be evaluated.
}
\description{
helper function for evaluating code from text
}
\details{
use "`r mean(c(2,3))`" to specify evaluating the \code{mean(c(2,3))}
}
\keyword{internal}

