% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{extract_feature}
\alias{extract_feature}
\title{extract and summarize spatial data}
\usage{
extract_feature(x, ...)
}
\arguments{
\item{x}{an object of class "Spatial" or a filepath that can be read in}

\item{\dots}{additional parameters passed to methods.
Including \code{out} a character vector of summary values}
}
\value{
a list according to names in spatial lookup tables for tag conversion
}
\description{
create metadata list from sp object
}
\examples{
library(sp)
Sr1 = Polygon(cbind(c(-89,-89.5,-89,-88.5,-89),c(42,42,44,44,42)))
Srs1 = Polygons(list(Sr1), "s1")
p = SpatialPolygons(list(Srs1), proj4string=CRS("+init=epsg:4326 +proj=longlat 
+datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"))
extract_feature(p)
}
