% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{feature_bbox}
\alias{feature_bbox}
\title{get the bounding box of the feature(s)}
\usage{
feature_bbox(sp)
}
\arguments{
\item{sp}{a spatial object from the \code{sp} package}
}
\value{
a list with \code{wbbox}, \code{ebbox}, \code{nbbox}, \code{sbbox} fields
}
\description{
Calculate the bounding box of the feature(s) in  lat/lon, formatted for metadata field entry
}
\examples{
library(sp)
Sr1 <- Polygon(cbind(c(-89,-89.5,-89,-88.5,-89),c(42,42,44,44,42)))
Srs1 <- Polygons(list(Sr1), "s1")
Sr2 <- Polygon(cbind(c(-105,-105.5,-106,-105.5,-105),c(31.4,32,34,34,32)))
Srs2 <- Polygons(list(Sr2), "s2")
p <- SpatialPolygons(list(Srs1, Srs2), proj4string=CRS("+proj=longlat +datum=WGS84"))
p.df <- SpatialPolygonsDataFrame(Sr = p, data = data.frame(x=c(0,40), y=c(30,300), z=c(0,0), row.names=c('s1','s2')))
library(rgdal)
writeOGR(p.df, 'inst/extdata/example_shapefile/','example_shapefile', driver = 'ESRI Shapefile', overwrite_layer = TRUE)
}

