% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{feature_bbox.Spatial}
\alias{feature_bbox.Spatial}
\title{get the bounding box of the feature(s)}
\usage{
\method{feature_bbox}{Spatial}(sp)
}
\arguments{
\item{sp}{a spatial object from the \code{sp} package}
}
\value{
a list with \code{wbbox}, \code{ebbox}, \code{nbbox}, \code{sbbox} fields
}
\description{
Calculate the bounding box of the feature(s) in  lat/lon, formatted for metadata field entry
}
\keyword{internal}

