% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{feature_count.Spatial}
\alias{feature_count.Spatial}
\title{get the feature count from a spatial object}
\usage{
\method{feature_count}{Spatial}(obj)
}
\arguments{
\item{sp}{a spatial object from the \code{sp} package}

\item{return}{a list with \code{feature-count} field}
}
\description{
Tally the number of features in a \code{sp} object
}
\keyword{internal}
