% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{feature_type.SpatialPolygons}
\alias{feature_type.SpatialPolygons}
\title{get the FGDC feature type of spatial object}
\usage{
\method{feature_type}{SpatialPolygons}(obj)
}
\arguments{
\item{obj}{a spatial object from the \code{sf} package or \code{sp} package}

\item{return}{a list with \code{feature-ref} and \code{feature-type} fields}
}
\description{
Extract the FGDC feature type from an \code{sp} object
}
\details{
not all \code{sf} or \code{sp} spatial classes are supported. 
If you run into one that you need, file an issue https://github.com/USGS-R/meddle/issues
}
\keyword{internal}
