% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-utils.R
\name{get_attrs}
\alias{get_attrs}
\title{read attributes from a recognized file type}
\usage{
get_attrs(filename)
}
\arguments{
\item{filename}{the full path for the file(s)}
}
\value{
the attributes as a character vector
}
\description{
return just the header names (attributes) for delimited files or shapefiles
}

