% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{get_feature_count}
\alias{get_feature_count}
\title{get the feature count from a spatial object}
\usage{
get_feature_count(sp)
}
\arguments{
\item{sp}{a spatial object from the \code{sp} package}

\item{return}{a list with \code{feature-count} field}
}
\description{
Tally the number of features in a \code{sp} object
}
\keyword{internal}

