% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-utils.R
\name{merge_attr_table}
\alias{merge_attr_table}
\title{merge two attribute tables}
\usage{
merge_attr_table(new.attrs, old.attrs)
}
\arguments{
\item{new.attrs}{an attribute table (data.frame)}

\item{old.attrs}{an attribute table (data.frame)}
}
\value{
a merged attribute table.
}
\description{
merge two attribute tables, assuming that the latter one
has more complete data, but the newer one has all of the applicable fields.
Fields in \code{old.attrs} that don't appear in \code{new.attrs} will be dropped.
Likewise, fields in \code{new.attrs} but not in \code{old.attrs} will retain
the values in \code{new.attrs}.
}

