% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_attributes.R
\name{parse_attributes}
\alias{parse_attributes}
\title{read and format attributes from delimited file}
\usage{
parse_attributes(filename)
}
\arguments{
\item{filename}{the name of the file for attributes}
}
\value{
a list of attributes and values for rendering
}
\description{
formats a file attribute table into a list appropriate for metadata rendering
}
\keyword{internal}

