% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_attrs}
\alias{read_attrs}
\title{read attributes from a recognized file}
\usage{
read_attrs(filename)
}
\arguments{
\item{filename}{the full path for the file(s)}
}
\value{
the attributes as a character vector
}
\description{
return just the header names (attributes) for delimited files or shapefiles
}

