% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{read data from supported file formats}
\usage{
read_data(filename, ...)
}
\arguments{
\item{filename}{a complete file path}

\item{\dots}{additional arguments passed to file-specific methods.}
}
\description{
read in data from filepath into R
}
