% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{rename_geometry}
\alias{rename_geometry}
\title{a fix for geom names that are different in `sf` objects
from https://gis.stackexchange.com/questions/386584/sf-geometry-column-naming-differences-r}
\usage{
rename_geometry(obj, name)
}
\description{
a fix for geom names that are different in `sf` objects
from https://gis.stackexchange.com/questions/386584/sf-geometry-column-naming-differences-r
}
\keyword{internal}
