% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render.list}
\alias{render.list}
\title{render from a list}
\usage{
\method{render}{list}(data, filename, ..., template)
}
\arguments{
\item{template}{a character or filepath}
}
\description{
render text to a file from a list
}
\details{
template can be missing, a filepath to a *.mustache file, or a string
}
\examples{
render(list('dogname'='fred','catname'='midred'),
   filename=NULL, list('dogname'='betty'), template="my dog's name is: {{dogname}}")
}
\keyword{internal}

