% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_attributes.R
\name{submissing}
\alias{submissing}
\title{a helper function for missing values}
\usage{
submissing(value, sub.val = "NA")
}
\arguments{
\item{value}{the actual value}

\item{sub.val}{the value to sub in if \code{value} is empty}
}
\value{
a modified value according to \code{sub.val}
}
\description{
creates a missing value filler if needed
}
\keyword{internal}

