% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_skeleton.R
\name{attribute_skeleton}
\alias{attribute_skeleton}
\title{create a skeleton file (or safely append to existing) for attributes}
\usage{
attribute_skeleton(object, attr.file, ...)
}
\arguments{
\item{object}{a regognized R object or vector of file paths that can be read in to a recognized object class}

\item{attr.file}{a file to write the attribute skeleton to}

\item{\dots}{additional parameters passed to methods}
}
\description{
create skeleton from object
}
\details{
if the attribute file (\code{attr.file}) doesn't exist, a skeleton formatted file is created.
If the file does exist, it is opened and any new fields are joined in. Existing content in
\code{attr.file} is not overwritten.
}

