% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{extract_feature}
\alias{extract_feature}
\title{extract and summarize spatial data}
\usage{
extract_feature(sp, out = c("bbox", "type", "count"))
}
\arguments{
\item{sp}{an object of class "Spatial"}

\item{out}{a character vector of summary values}
}
\value{
a list according to names in spatial lookup tables for tag conversion
}
\description{
create metadata list from sp object
}

