% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{feature_states}
\alias{feature_states}
\title{get the states that features overlap with}
\usage{
feature_states(sf)
}
\arguments{
\item{sp}{an object of class \code{Spatial}}
}
\value{
a list with \code{states} field
}
\description{
summarize the spatial extent of data relative to overlap w/ US states
}
\examples{
library(sf)
sfc = st_sfc(st_point(c(-89, 45)), st_point(c(-109, 42)), crs = 4326)
sf = st_sf(data.frame(val = c('WI_point', 'WY_point'), geom=sfc))
feature_states(sf)
}
