% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial.R
\name{feature_type.SpatialPolygons}
\alias{feature_type.SpatialPolygons}
\title{get the FGDC feature type of spatial object}
\usage{
\method{feature_type}{SpatialPolygons}(sp)
}
\arguments{
\item{sp}{a spatial object from the \code{sp} package}

\item{return}{a list with \code{feature-type} field}
}
\description{
Extract the FGDC feature type from an \code{sp} object
}
\details{
only classes SpatialPointsDataFrame and SpatialPolygonsDataFrame classes are currently supported
}
\keyword{internal}
