% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{render xml from data and template}
\usage{
render(data, filename, ...)
}
\arguments{
\item{data}{filepath or list}

\item{filename}{name of file to write metadata to}

\item{\dots}{additional lists to include or other arguments passed to methods
(e.g., \code{template="metadata.mustache"})}
}
\description{
create metadata file from input data or yaml file
}

